Tied most loose ends
1. Borrow and Return actions from the customers controller now update
   the count of the movies.
2. Deleting a movie now updates the ledger by removing all entries from
   the ledger corresponding to that movie.
3. Added an action for customers to delete their account from the
   account settings view.
4. Deleting a customer (either from the customers view of the admin
   controller or from the delete account option in the customer's
   account settings) now updates the ledger and updates the counts of
   movies if any were borrowed by the customer and not returned before
   the account was deleted.

All of this was made possible through reuse of code wherever applicable
by creating an abstract base controller class (with void virtual methods)
inherited by the other controllers which can then call its functions.

All that's left to be implemented are form validations, alerts and
abstracting the url to improve security (even if it is to do so
marginally).