Final commit
This will be my last commit to this project for a while. If I ever
decide to continue working on this project I'll probably add pagination
to the the landing page of customers controller for the mnovies section,
add confirm password textbox to edit/create actions on customer
instances and add access control to prevent navigation from the address
bar.